Make utils a dedicated package that will pip installable later and rename utils.py into parsing.py
