--- __init__.py
@@ -1,3 +1,0 @@
-import schemas
-
-__all__ = ["schemas"]
