+++ examples/mlee.py
@@ -23,7 +23,7 @@ import datasets
 from typing import List
 from dataclasses import dataclass
 
-from biodatasets import utils
+from utils import parsing
 
 _DATASETNAME = "mlee"
 _SOURCE_VIEW_NAME = "source"
@@ -325,14 +325,14 @@ class MLEE(datasets.GeneratorBasedBuilder):
         if self.config.schema == "source":
             txt_files = list(data_files.glob("*txt"))
             for guid, txt_file in enumerate(txt_files):
-                example = utils.parse_brat_file(txt_file)
+                example = parsing.parse_brat_file(txt_file)
                 example["id"] = str(guid)
                 yield guid, example
         elif self.config.schema == "bigbio_kb":
             txt_files = list(data_files.glob("*txt"))
             for guid, txt_file in enumerate(txt_files):
-                example = utils.brat_parse_to_bigbio_kb(
-                    utils.parse_brat_file(txt_file),
+                example = parsing.brat_parse_to_bigbio_kb(
+                    parsing.parse_brat_file(txt_file),
                     entity_types=self._ENTITY_TYPES
                 )
                 example["id"] = str(guid)
